+++ src/lib.rs
@@ -1,8 +1,10 @@
 mod config;
+mod inputs;
 mod language_name;
 pub mod loader;
 mod range_union;
 pub mod searches;
+pub mod main_search;
 
 pub use config::{Config, QueryCompiler};
 pub use language_name::LanguageName;